requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
google-generativeai==0.5.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
pymongo==4.6.2
dramatiq[redis]
redis
requests